Create README.md
This app acts as morse code translator, turning your text into morse code. Then at the push of a button you are able to transform the morse code into beeps or flashes. The idea behind this app was to offer a long range means of communication without the necessity of phone signal or wifi, a use case would be if you got lost in the woods and needed to signal for help. My app can flash messages you created or can be selected from pre-selected codes in the app. An overall handy app, and potentially life-saving!